#35: Extract GridKeyboardNavService from Grid component
